Ajout de l'antialiasing sur le dessin

+++ src/c/ControleurVueDessin.java
@@ -1,11 +1,7 @@
 package c;
 
 import java.awt.Color;
-
-import m.Orientation;
-import m.Strategie;
 import v_vues.Vue_Dessin;
-import v_vues.Vue_Terrain;
 
 public class ControleurVueDessin {
 	private ControleurPrincipal cp;
@@ -65,13 +61,13 @@ public class ControleurVueDessin {
 	public void jeton(m.Position _pos)
 	{
 		Color c = this.cp.getP().getCoulE(0);
-		int t = (int)(0.75*this.tc);
-		int posX = this.tc*(_pos.getPosX()+1) + this.tc/2 - t/2;
-		int posY = this.tc*(_pos.getPosY()+1) + this.tc/2 - t/2;
+		float t = (float)0.75*this.tc;
+		float posX = this.tc*(_pos.getPosX()+1) + this.tc/2 - t/2;
+		float posY = this.tc*(_pos.getPosY()+1) + this.tc/2 - t/2;
 		this.vd.affJeton(posX, posY, _pos.getOrt(), t, c);
 	}
 	
-	public int getTC()
+	public float getTC()
 	{
 		return this.tc;
 	}
